requirements.txt
pytesseract
Pillow
transformers
pymongo